Refactor: Break down HomeScreen into multiple widgets
- Extracted UI sections of HomeScreen into individual, more manageable widgets under `lib/screens/home/widgets/`.
- Each new widget now handles its own content display and shimmer loading state.
- Cleaned up HomeScreen by removing the extracted build methods and unused imports/variables.
- This improves code organization, readability, and maintainability of the home screen.
