+++ lib/screens/home/widgets/.placeholder
@@ -1,0 +1,2 @@
+# This is a placeholder file to create the directory.
+# It can be removed later if desired.
